Aligned chat-score-wrapper to game
It took some finagling, but I've got the actual game window centered on the page and the chat/leaderboard bits to the right of it. Added a game-wrapper div and a content-wrapper div, messed with so many different position and display properties that I don't even remember which ones I changed.